Guile Scheme replacement for interpolate.sh
The City of Austin skips days of reporting every weekend. I have been manually
linearly interpolating those days, but this is annoying and I'd like to have
the script do it for me. Unfortunately, the awk version is already at its complexity
limits. I kind of wanted to do this in a Lisp, so I am doing it in Scheme.

At this point, the date, cases, arith, and integral arithsmoothed numbers match
the old output, but non-integral arithsmoothed numbers are being output as
fractions, so they are harder to compare. Also, we do not actually yet do the
linear interpolation.