Remove interpolated data
The script now does this for us, so there is no reason to keep the
manual interpolations. The orders of the added numbers may differ
between the old manual interpolations and the new ones. In particular,
when interpolating manually I always attempted to have the day that
differed come last, whether it was larger or smaller than the other
days involved. The script will always have larger days first.