Code for automatically pulling from the data set
Recently found a link to this from Reddit. It was not easy to find: I'd previously looked for
it with no success, but now that it's here, expect a cron job.

Data here: https://atc-covid19data-austin.hub.arcgis.com/

This does introduce a dependency on jq for parsing the results JSON returned from ARCGIS.